% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbQuoteLiteralMethodPool.R
\name{dbQuoteLiteral,Pool-method}
\alias{dbQuoteLiteral,Pool-method}
\title{dbQuoteLiteral}
\usage{
\S4method{dbQuoteLiteral}{Pool}(conn, x, ...)
}
\description{
dbQuoteLiteral
}
\details{
As of 2020-05-07, this is necessary due to an incompatiblity
between packages `pool` (v 0.1.4.3) and `glue` (v >= 1.3.2).
In v1.3.2, `glue` started using `dbQuoteLiteral`, which is
currently not offered by `pool`, so creating it here.
}
