% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_utils.R
\name{pool_add_column}
\alias{pool_add_column}
\title{Pool Show}
\usage{
pool_add_column(
  pool,
  table,
  column,
  type,
  not_null = FALSE,
  .before = NULL,
  .after = NULL
)
}
\description{
show the way to create TABLE
}
