% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_utils.R
\name{pool_describe}
\alias{pool_describe}
\title{Pool Describe}
\usage{
pool_describe(pool, table)
}
\description{
used to get fields information in database
}
\note{
using for MySQL Database
}
\examples{
pool_describe(lime, table = "branchs_intraday")
}
