% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_utils.R
\name{pool_drop_key}
\alias{pool_drop_key}
\title{Pool Drop Key}
\usage{
pool_drop_key(pool, table, key = c("primary", "foreign"), constraint = NULL)
}
\description{
drop key (foreign or primary) of a table in databsse
}
