% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_execute_many.R
\name{pool_execute_many}
\alias{pool_execute_many}
\title{Execute Many}
\usage{
pool_execute_many(pool, statement, data, ...)
}
\description{
implement execute many information by using only one execute
}
\references{

}
