% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_utils.R
\name{pool_grant}
\alias{pool_grant}
\title{Pool Grant}
\usage{
pool_grant(pool, privilege, object, user, grant_options = FALSE)
}
\arguments{
\item{privilege}{privilege of user in objects from pool user}

\item{object}{name of objects in database, included: TABLE, VIEW, STORED PROC and SEQUENCE.}

\item{grant_options}{WITH_GRANT OPTIONS (default FALSE) give permissions for other users from user}
}
\description{
grant privilege for user
}
\references{
\href{MYSQL Privilege}{https://dev.mysql.com/doc/refman/8.0/en/grant.html}
\href{}{https://thuthuat.taimienphi.vn/grant-revoke-trong-sql-33413n.aspx}
}
