% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_utils.R
\name{pool_head}
\alias{pool_head}
\title{Pool Show}
\usage{
pool_head(pool, table, n = 5)
}
\description{
show the way to create TABLE
}
\examples{
pool_call()
tables <- DBI::dbListTables(lime)
pool_show(lime, table = tables[1])
}
