% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_utils.R
\name{pool_nrow}
\alias{pool_nrow}
\title{Pool Nrow}
\usage{
pool_nrow(pool, table)
}
\description{
count number of rows in specific table
}
\examples{
pool_call()
tables <- DBI::dbListTables(lime)
pool_nrow(lime, table = tables[1])
}
