% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_method.R
\name{pool_query}
\alias{pool_query}
\title{Pool Query}
\usage{
pool_query(
  pool,
  statement,
  .parse = TRUE,
  .parse_funs = tidytable::as_tidytable,
  ...
)
}
\arguments{
\item{pool}{pool object, created by `pool::dbPool`}

\item{statement}{SQL statement send to database}

\item{...}{Other parameters passed on to methods align with `DBI` package}
}
\description{
method of query to database managed by pool
}
