% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_load.R
\name{pool_set}
\alias{pool_set}
\title{Pool Set}
\usage{
pool_set(directory = "/var/lib/mysql")
}
\description{
set environment for using pool load wit will connect to target host
}
\details{
Here are where the files are installed on the system:
All configuration files (like my.cnf) are under /etc/mysql
All binaries, libraries, headers, etc., are under /usr/bin and /usr/sbin
The data directory is under /var/lib/mysql

4. (Optional) Set up SSH Agent to store the keys to avoid having to re-enter passphrase at every login
Enter the following commands to start the agent and add the private SSH key
ssh-agent $BASH
ssh-add ~/.ssh/id_rsa
Type in your key’s current passphrase when asked. If you saved the private key somewhere other than the default location and name, you’ll have to specify it when adding the key.
}
\references{
\url{https://upcloud.com/community/tutorials/use-ssh-keys-authentication/}
\url{https://mariadb.com/kb/en/configuring-mariadb-for-remote-client-access/}{Set Data}
\url{https://community.spiceworks.com/topic/484926-mysql-error-ignoring-query-to-other-database}
}
