% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_utils.R
\name{pool_size}
\alias{pool_size}
\title{Pool Size}
\usage{
pool_size(pool, name, size_of = c("table", "schema"), unit = "MB")
}
\description{
drop key (foreign or primary) of a table in databsse
}
\references{
\href{Size of table database}{https://www.a2hosting.com/kb/developer-corner/mysql/determining-the-size-of-mysql-databases-and-tables}
}
