% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_utils.R
\name{pool_version}
\alias{pool_version}
\title{Pool Version}
\usage{
pool_version(pool)
}
\description{
get version of database
}
\examples{
pool_version(pool = lime)
}
\references{
MySQL :: MySQL 8.0 Reference Manual :: 12.15 Information Functions
}
