% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_load.R
\name{remote_file_exists}
\alias{remote_file_exists}
\title{Remote File Exists}
\usage{
remote_file_exists(path, user, host, port = 22, password = NULL)
}
\description{
validate file is exists from a remote file
}
