% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_load.R
\name{remote_transfer}
\alias{remote_transfer}
\title{Remote Transfer}
\usage{
remote_transfer(
  path,
  new_path,
  server = NULL,
  port = 22,
  user,
  password,
  .message = TRUE,
  method = "scp"
)
}
\description{
Transfer Using SCP
}
\details{
Syntax of transfer:
<scp [OPTION] [user@]SRC_HOST:]file1 [user@]DEST_HOST:]file2>

chown -R username directory

Common Error
1. Can't create regular file: ... Permission Deninded

2. Setting User with mode 666 / 777 by root user

On Unix, if you need LOAD DATA to read from a pipe, you can use the following technique (the example loads a listing of the / directory into the table db1.t1):

# mkfifo /mysql/data/db1/ls.dat
# chmod 666 /mysql/data/db1/ls.dat
# find / -ls > /mysql/data/db1/ls.dat &
# mysql -e "LOAD DATA INFILE 'ls.dat' INTO TABLE t1" db1

It need packages sshpass in linux to work with, you can install it in cli by using: apt install sshpass
}
\references{
\url{https://unix.stackexchange.com/questions/466625/cp-cannot-create-regular-file-permission-denied}{}
\url{https://unix.stackexchange.com/questions/52634/error-using-scp-not-a-regular-file}{Unix Exchange}
}
